+++ src/main/java/com/sphereex/cases/proxy/transaction/readonly/SetReadOnly.java
@@ -4,6 +4,8 @@ import com.sphereex.cases.ProxyBaseTest;
 import com.sphereex.core.AutoTest;
 import com.sphereex.core.CaseInfo;
 import com.sphereex.core.DBType;
+import com.sphereex.core.Status;
+import lombok.Getter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -14,43 +16,55 @@ import java.sql.Statement;
 
 @AutoTest
 public final class SetReadOnly extends ProxyBaseTest {
+    
+    @Getter
+    private final DBType dbType = DBType.MYSQL;
 
     private static final Logger logger = LoggerFactory.getLogger(SetReadOnly.class);
     
-    public SetReadOnly() {
-        super(DBType.MYSQL);
-    }
-    
     @Override
-    public void pre() throws Exception {
-        Connection conn = getAutodataSource().getConnection();
-        Statement stmt;
-        Statement stmt1;
-        Statement stmt2;
-        stmt = conn.createStatement();
-        stmt.executeUpdate("drop table if exists account;");
-        stmt.close();
-        stmt1 = conn.createStatement();
-        stmt1.executeUpdate("create table account(id int primary key , balance int not null);");
-        stmt1.close();
-        stmt2 = conn.createStatement();
-        stmt2.executeUpdate("insert into account(id,balance) values (1,0),(2,100);");
-        stmt2.close();
-        conn.close();
+    public Status pre() {
+        Status s = super.pre();
+        if (!s.isSuccess()) {
+            return s;
+        }
+        try {
+            Connection conn = getAutoDataSource().getConnection();
+            Statement stmt;
+            Statement stmt1;
+            Statement stmt2;
+            stmt = conn.createStatement();
+            stmt.executeUpdate("drop table if exists account;");
+            stmt1 = conn.createStatement();
+            stmt1.executeUpdate("create table account(id int primary key , balance int not null);");
+            stmt2 = conn.createStatement();
+            stmt2.executeUpdate("insert into account(id,balance) values (1,0),(2,100);");
+        } catch (SQLException e) {
+            e.printStackTrace();
+            return new Status(false, e.getMessage());
+        }
+        return new Status(true, "");
     }
 
     @Override
-    public boolean run() throws Exception {
-        boolean r1 = step1();
-        boolean r2 = step2();
-        if (!r1 || !r2) {
-            return false;
+    public Status run() {
+        try {
+            step1();
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new Status(false, e.getMessage());
         }
-        return true;
+        try {
+            step2();
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new Status(false, e.getMessage());
+        }
+        return new Status(true, "");
     }
     
-    private boolean step1() throws Exception {
-        Connection conn = getAutodataSource().getConnection();
+    private void step1() throws Exception {
+        Connection conn = getAutoDataSource().getConnection();
         conn.setReadOnly(true);
         Statement statement1 = conn.createStatement();
         ResultSet rs = statement1.executeQuery("select * from account;");
@@ -60,13 +74,13 @@ public final class SetReadOnly extends ProxyBaseTest {
             if (id == 1) {
                 if (balance != 0) {
                     logger.error("balance is {}, should be 0", balance);
-                    return false;
+                    throw new Exception(String.format("balance is {}, should be 0", balance));
                 }
             }
             if (id == 2) {
                 if (balance != 100) {
                     logger.error("balance is {}, should be 100", balance);
-                    return false;
+                    throw new Exception(String.format("balance is {}, should be 100", balance));
                 }
             }
         }
@@ -74,18 +88,15 @@ public final class SetReadOnly extends ProxyBaseTest {
         try {
             statement2.execute("update account set balance=100 where id=2;");
             logger.error("update run success, should failed");
-            return false;
+            throw new Exception("update run success, should failed");
         } catch (SQLException e) {
             logger.info("update failed for expect");
+            throw new Exception("update failed for expect");
         }
-        statement1.close();
-        statement2.close();
-        conn.close();
-        return true;
     }
 
-    private boolean step2() throws Exception {
-        Connection conn = getAutodataSource().getConnection();
+    private void step2() throws Exception {
+        Connection conn = getAutoDataSource().getConnection();
         Statement statement1 = conn.createStatement();
         ResultSet rs = statement1.executeQuery("select * from account");
         while (rs.next()) {
@@ -94,13 +105,13 @@ public final class SetReadOnly extends ProxyBaseTest {
             if (id == 1) {
                 if (balance != 0) {
                     logger.error("balance is {}, should be 0", balance);
-                    return false;
+                    throw new Exception(String.format("balance is {}, should be 0", balance));
                 }
             }
             if (id == 2) {
                 if (balance != 100) {
                     logger.error("balance is {}, should be 100", balance);
-                    return false;
+                    throw new Exception(String.format("balance is {}, should be 100", balance));
                 }
             }
         }
@@ -111,18 +122,17 @@ public final class SetReadOnly extends ProxyBaseTest {
         ResultSet r3 = statement3.executeQuery("select * from account where id=2");
         if (!r3.next()) {
             logger.error("update run failed, should success");
-            return false;
+            throw new Exception("update run failed, should success");
         }
         int balanceend = r3.getInt("balance");
         if (balanceend != 101) {
             logger.error("update run failed, should success");
-            return false;
+            throw new Exception("update run failed, should success");
         }
-        return true;
     }
 
     @Override
-    public void initCaseInfo() {
+    public void initCase() {
         String name = "SetReadOnly";
         String feature = "proxy-transaction";
         String tag = "MySQL";
@@ -132,6 +142,7 @@ public final class SetReadOnly extends ProxyBaseTest {
                 "3. session B run 'update' successful";
         CaseInfo caseInfo = new CaseInfo(name, feature, tag, message);
         setCaseInfo(caseInfo);
+        super.initCase();
     }
 }
