+++ src/main/java/com/sphereex/cases/traffic/TrafficSetReadOnlyTest.java
@@ -2,6 +2,7 @@ package com.sphereex.cases.traffic;
 
 import com.sphereex.core.AutoTest;
 import com.sphereex.core.CaseInfo;
+import com.sphereex.core.Status;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -20,30 +21,49 @@ public final class TrafficSetReadOnlyTest extends TrafficBaseTest {
     }
     
     @Override
-    public void pre() throws Exception {
-        Connection conn = getAutoDataSource().getConnection();
-        Statement stmt;
-        Statement stmt1;
-        stmt = conn.createStatement();
-        stmt.executeUpdate("delete from t_order;");
-        stmt.close();
-        stmt1 = conn.createStatement();
-        stmt1.executeUpdate("INSERT INTO t_order(user_id, content, creation_date, order_id) VALUES(1, 'test1', NOW(), 1), (1, 'test2', NOW(),2);");
-        stmt1.close();
-        conn.close();
+    public Status pre() {
+        Status s = super.pre();
+        if (!s.isSuccess()) {
+            return s;
+        }
+        try {
+            Connection conn = getAutoDataSource().getConnection();
+            Statement stmt;
+            Statement stmt1;
+            stmt = conn.createStatement();
+            stmt.executeUpdate("delete from t_order;");
+            stmt.close();
+            stmt1 = conn.createStatement();
+            stmt1.executeUpdate("INSERT INTO t_order(user_id, content, creation_date, order_id) VALUES(1, 'test1', NOW(), 1), (1, 'test2', NOW(),2);");
+            stmt1.close();
+            conn.close();
+        } catch (SQLException e) {
+            e.printStackTrace();
+            return new Status(false, e.getMessage());
+        }
+        return new Status(true, "");
+        
     }
     
     @Override
-    public boolean run() throws Exception {
-        boolean r1 = step1();
-        boolean r2 = step2();
-        if (!r1 || !r2) {
-            return false;
+    public Status run() {
+        try {
+            step1();
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new Status(false, e.getMessage());
         }
-        return true;
+    
+        try {
+            step2();
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new Status(false, e.getMessage());
+        }
+        return new Status(true, "");
     }
     
-    private boolean step1() throws Exception {
+    private void step1() throws Exception {
         Connection conn = getAutoDataSource().getConnection();
         conn.setReadOnly(true);
         Statement statement1 = conn.createStatement();
@@ -54,13 +74,13 @@ public final class TrafficSetReadOnlyTest extends TrafficBaseTest {
             if (id == 2) {
                 if (!"test2".equals(content)) {
                     logger.error("content is {}, should be test2", content);
-                    return false;
+                    throw new Exception(String.format("content is {}, should be test2", content));
                 }
             }
             if (id == 1) {
                 if (!"test1".equals(content)) {
                     logger.error("content is {}, should be test1", content);
-                    return false;
+                    throw new Exception(String.format("content is {}, should be test1", content));
                 }
             }
         }
@@ -68,17 +88,16 @@ public final class TrafficSetReadOnlyTest extends TrafficBaseTest {
         try {
             statement2.execute("update t_order set content='test3' where order_id=2;");
             logger.error("update run success, should failed");
-            return false;
+            throw new Exception("update run success, should failed");
         } catch (SQLException e) {
             logger.info("update failed for expect");
         }
         statement1.close();
         statement2.close();
         conn.close();
-        return true;
     }
     
-    private boolean step2() throws Exception {
+    private void step2() throws Exception {
         Connection conn = getAutoDataSource().getConnection();
         Statement statement1 = conn.createStatement();
         ResultSet rs = statement1.executeQuery("select * from t_order");
@@ -88,13 +107,13 @@ public final class TrafficSetReadOnlyTest extends TrafficBaseTest {
             if (id == 2) {
                 if (!"test2".equals(content)) {
                     logger.error("content is {}, should be test2", content);
-                    return false;
+                    throw new Exception(String.format("content is {}, should be test2", content));
                 }
             }
             if (id == 1) {
                 if (!"test1".equals(content)) {
                     logger.error("content is {}, should be test1", content);
-                    return false;
+                    throw new Exception(String.format("content is {}, should be test1", content));
                 }
             }
         }
@@ -105,27 +124,32 @@ public final class TrafficSetReadOnlyTest extends TrafficBaseTest {
         ResultSet r3 = statement3.executeQuery("select * from t_order where order_id=2");
         if (!r3.next()) {
             logger.error("update run failed, should success");
-            return false;
+            throw new Exception("update run failed, should success");
         }
         String content = r3.getString("content");
         if (!"test3".equals(content)) {
             logger.error("update run failed, should success");
-            return false;
+            throw new Exception("update run failed, should success");
         }
-        return true;
     }
     
     @Override
-    public void end() throws SQLException {
-        Connection conn = getAutoDataSource().getConnection();
-        Statement statement = conn.createStatement();
-        statement.executeUpdate("delete from t_order;");
-        statement.close();
-        conn.close();
+    public Status end() {
+        try {
+            Connection conn = getAutoDataSource().getConnection();
+            Statement statement = conn.createStatement();
+            statement.executeUpdate("delete from t_order;");
+            statement.close();
+            conn.close();
+        } catch (SQLException e) {
+            e.printStackTrace();
+            return new Status(false, e.getMessage());
+        }
+        return new Status(true, "");
     }
     
     @Override
-    public void initCaseInfo() {
+    public void initCase() {
         String name = "TrafficSetReadOnlyTest";
         String feature = "traffic-transaction";
         String tag = "opengauss";
